+++ src/misc.js
@@ -1,5 +1,5 @@
 import * as MODULE from "../MaterialKeys.js";
-import {launchpad,soundboard,macroBoard} from "../MaterialKeys.js";
+import {launchpad,soundboard,macroBoard,playlistControl} from "../MaterialKeys.js";
 
 export class playlistConfigForm extends FormApplication {
     constructor(data, options) {
@@ -24,26 +24,37 @@ export class playlistConfigForm extends FormApplication {
      * Provide data to the template
      */
     getData() {
-        const selectedPlaylists = game.settings.get(MODULE.moduleName,'selectedPlaylists');
-        let playlistData = {};
-
+        let settings = game.settings.get(MODULE.moduleName,'playlists');
+        let selectedPlaylists = settings.selectedPlaylist;
+        if (selectedPlaylists == undefined) selectedPlaylists = [];
+        let selectedPlaylistMode = settings.playlistMode;
+        if (selectedPlaylistMode == undefined) selectedPlaylistMode = [];
+        let playMode = settings.playMode;
+        if (playMode == undefined) playMode = 0;
+        let playlistData = [];
+        
         for (let i=0; i<8; i++){
-           // const playlist = MODULE.getFromJSONArray(selectedPlaylists,i);
-            const dataThis = {
+            if (selectedPlaylists[i] == undefined) selectedPlaylists[i] = 'none';
+            if (selectedPlaylistMode[i] == undefined) selectedPlaylistMode[i] = 0;
+            let dataThis = {
                 iteration: i+1,
                 playlist: selectedPlaylists[i],
+                playlistMode: selectedPlaylistMode[i],
                 playlists: game.playlists.entities
             }
-            MODULE.setToJSONArray(playlistData,i,dataThis);
+            playlistData.push(dataThis);
         }
-    
-        if (!this.data && selectedPlaylists) {
-            this.data = selectedPlaylists;
+
+        this.data = {
+            playMode: playMode,
+            selectedPlaylist: selectedPlaylists,
+            playlistMode: selectedPlaylistMode
         }
+
         return {
             playlists: game.playlists.entities,
             playlistData: playlistData,
-            playMethod: game.settings.get(MODULE.moduleName,'playlistMethod')
+            playMode: playMode
         } 
     }
 
@@ -53,13 +64,42 @@ export class playlistConfigForm extends FormApplication {
      * @param {*} formData 
      */
     async _updateObject(event, formData) {
-        await game.settings.set(MODULE.moduleName,'selectedPlaylists', formData["selectedPlaylist"]);
-        await game.settings.set(MODULE.moduleName,'playlistMethod',formData["playMethod"]);
+       // await game.settings.set(MODULE.moduleName,'selectedPlaylists', formData["selectedPlaylist"]);
+       // await game.settings.set(MODULE.moduleName,'playlistMethod',formData["playMethod"]);
     }
 
     activateListeners(html) {
         super.activateListeners(html);  
+        const playMode = html.find("select[name='playMode']");
+        const selectedPlaylist = html.find("select[name='selectedPlaylist']");
+        const playlistMode = html.find("select[name='playlistMode']");
+
+        playMode.on("change", event => {
+            this.data.playMode=event.target.value;
+            this.updateSettings(this.data);
+        });
+
+        selectedPlaylist.on("change", event => {
+            let id = event.target.id.replace('playlist','');
+            this.data.selectedPlaylist[id-1]=event.target.value;
+            this.updateSettings(this.data);
+        });
+
+        playlistMode.on("change", event => {
+            let id = event.target.id.replace('playlistMode','');
+            this.data.playlistMode[id-1]=event.target.value;
+            this.updateSettings(this.data);
+        });
     }
+    async updateSettings(settings){
+        await game.settings.set(MODULE.moduleName,'playlists', settings);
+        if (MODULE.enableModule) {
+            playlistControl.playlistUpdate();
+            playlistControl.volumeUpdate();
+        }
+        //this.render();
+    }
+    
 }
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -68,8 +108,6 @@ export class soundboardConfigForm extends FormApplication {
     constructor(data, options) {
         super(data, options);
         this.data = data;
-        this.playlist;
-        this.updatePlaylist = false;
     }
 
     /**
@@ -95,58 +133,85 @@ export class soundboardConfigForm extends FormApplication {
      * Provide data to the template
      */
     getData() {
-        let playlistId = game.settings.get(MODULE.moduleName,'soundboardSettings').playlist;
-        if (this.updatePlaylist) playlistId = this.playlist;
-        this.updatePlaylist = false;
-        let playlist = 'none';
-        let sounds = [];
-        if (playlistId != undefined){
-            playlist = game.playlists.entities.find(p => p._id == playlistId);
-            if (playlist != undefined) sounds = playlist.sounds;
-            else playlist = 'none';
+        let settings = game.settings.get(MODULE.moduleName,'soundboardSettings');
+        let playlists = [];
+        playlists.push({id:"none",name:game.i18n.localize("MaterialKeys.None")});
+        playlists.push({id:"FP",name:game.i18n.localize("MaterialKeys.FilePicker")})
+        for (let i=0; i<game.playlists.entities.length; i++){
+            playlists.push({id:game.playlists.entities[i]._id,name:game.playlists.entities[i].name});
         }
-        let selectedSounds = game.settings.get(MODULE.moduleName,'soundboardSettings').sounds;
-        let colorOn = game.settings.get(MODULE.moduleName,'soundboardSettings').colorOn;
-        let colorOff = game.settings.get(MODULE.moduleName,'soundboardSettings').colorOff;
-        let mode = game.settings.get(MODULE.moduleName,'soundboardSettings').mode;
-        let toggle = game.settings.get(MODULE.moduleName,'soundboardSettings').toggle;
-        let volume = game.settings.get(MODULE.moduleName,'soundboardSettings').volume;
+        if (settings.sounds == undefined) settings.sounds = [];
+        if (settings.colorOn == undefined) settings.colorOn = [];
+        if (settings.colorOff == undefined) settings.colorOff = [];
+        if (settings.mode == undefined) settings.mode = [];
+        if (settings.volume == undefined) settings.volume = [];
+        if (settings.name == undefined) settings.name = [];
+        if (settings.selectedPlaylists == undefined) settings.selectedPlaylists = [];
+        if (settings.src == undefined) settings.src = [];
+        if (settings.toggle == undefined) settings.toggle = [];
 
-        if (selectedSounds == undefined) selectedSounds = [];
-        if (colorOn == undefined) colorOn = [];
-        if (colorOff == undefined) colorOff = [];
-        if (mode == undefined) mode = [];
-        if (toggle == undefined) toggle = [];
-        let soundData = {};
+        let soundData = [];
+        let iteration = 0;
 
         for (let j=0; j<8; j++){
-            let soundsThis = {};
+            let soundsThis = [];
             for (let i=0; i<8; i++){
-                if (volume == undefined) volume = 50;
+                let selectedPlaylist;
+                let sounds = [];
+                if (settings.volume[iteration] == undefined) settings.volume[iteration] = 50;
+                if (settings.selectedPlaylists[iteration]==undefined) {
+                    if (settings.sounds[iteration] != undefined && settings.sounds[iteration] != "") {
+                        selectedPlaylist = game.settings.get(MODULE.moduleName,'soundboardSettings').playlist;
+                        const pl = game.playlists.entities.find(p => p._id == selectedPlaylist);
+                        selectedPlaylist = pl._id;
+                        settings.selectedPlaylists[iteration]=selectedPlaylist;
+                        this.updateSettings(settings);
+                        sounds = pl.sounds;
+                    }
+                    else
+                        selectedPlaylist = 'none';
+                }
+                else if (settings.selectedPlaylists[iteration] == 'none' || settings.selectedPlaylists[iteration] == '') selectedPlaylist = 'none';
+                else if (settings.selectedPlaylists[iteration] == 'FP') selectedPlaylist = 'FP';
+                else {
+                    const pl = game.playlists.entities.find(p => p._id == settings.selectedPlaylists[iteration]);
+                    selectedPlaylist = pl._id;
+                    sounds = pl.sounds;
+                }
+                let styleSS = "";
+                let styleFP ="display:none";
+                if (selectedPlaylist == 'FP') {
+                    styleSS = 'display:none';
+                    styleFP = ''
+                }
                 
                 const dataThis = {
-                    iteration: 10*(8-j)+i+1,
-                    sound: selectedSounds[j*8+i],
+                    iteration: iteration+1,
+                    playlists: playlists,
+                    selectedPlaylist: selectedPlaylist,
+                    sound: settings.sounds[iteration],
                     sounds: sounds,
-                    colorOn: colorOn[j*8+i],
-                    colorOff: colorOff[j*8+i],
-                    mode: mode[j*8+i],
-                    toggle: toggle[j*8+i],
-                    volume: volume[j*8+i]
+                    srcPath: settings.src[iteration],
+                    colorOn: settings.colorOn[iteration],
+                    colorOff: settings.colorOff[iteration],
+                    mode: settings.mode[iteration],
+                    volume: settings.volume[iteration],
+                    name: settings.name[iteration],
+                    styleSS: styleSS,
+                    styleFP: styleFP,
+                    toggle: settings.toggle[iteration]
                 }
-                MODULE.setToJSONArray(soundsThis,i,dataThis);
+                soundsThis.push(dataThis);
+                iteration++;
             }
             const data = {
                 dataThis: soundsThis,
             };
-            MODULE.setToJSONArray(soundData,j,data);
+            soundData.push(data);
         }
+        
         return {
-            playlists: game.playlists.entities,
-            playlist: playlistId,
-            sounds: sounds,
-            selectedSound81: selectedSounds.a,
-            soundData: soundData,
+            soundData: soundData
         } 
     }
 
@@ -156,33 +221,73 @@ export class soundboardConfigForm extends FormApplication {
      * @param {*} formData 
      */
     async _updateObject(event, formData) {
-        await game.settings.set(MODULE.moduleName,'soundboardSettings',{
-            playlist: formData["playlist"],
-            sounds: formData["sounds"],
-            colorOn: formData["colorOn"],
-            colorOff: formData["colorOff"],
-            mode: formData["mode"],
-            toggle: formData["toggle"],
-            volume: formData["volume"]
-        });
-        soundboard.update();
+  
     }
 
     async activateListeners(html) {
         super.activateListeners(html);
-        const colorPickerOn = html.find("button[name='colorPickerOn']");
-        const colorPickerOff = html.find("button[name='colorPickerOff']");
+        const nameField = html.find("input[name='namebox']");
         const playlistSelect = html.find("select[name='playlist']");
-        const volumeSlider = html.find("input[name='volume']");
         const soundSelect = html.find("select[name='sounds']");
+        const soundFP = html.find("input[name2='soundSrc']");
+        const colorToggle = html.find("select[name='toggle'");
+        const colorPickerOn = html.find("button[name='colorPickerOn']");
+        const colorPickerOnNr = html.find("input[name='colorOn']");
+        const colorPickerOff = html.find("button[name='colorPickerOff']");
+        const colorPickerOffNr = html.find("input[name='colorOff']");
+        const playMode = html.find("select[name='mode']");
+        const volumeSlider = html.find("input[name='volume']");
+        const clearAll = html.find("button[name='clearAll']");
+
+        nameField.on("change",event => {
+            let id = event.target.id.replace('name','')-1;
+            let settings = game.settings.get(MODULE.moduleName,'soundboardSettings');
+            settings.name[id]=event.target.value;
+            this.updateSettings(settings);
+        });
+
+        playlistSelect.on("change", event => {
+            let id = event.target.id.replace('playlists','')-1;
+            let settings = game.settings.get(MODULE.moduleName,'soundboardSettings');
+            settings.selectedPlaylists[id]=event.target.value;
+            this.updateSettings(settings,true);
+        });
+        
+        soundSelect.on("change",event => {
+            let id = event.target.id.replace('soundSelect','')-1;
+            let settings = game.settings.get(MODULE.moduleName,'soundboardSettings');
+            settings.sounds[id]=event.target.value;
+            this.updateSettings(settings);
+        });
+        
+        soundFP.on("change",event => {
+            let id = event.target.id.replace('srcPath','')-1;
+            let settings = game.settings.get(MODULE.moduleName,'soundboardSettings');
+            settings.src[id]=event.target.value;
+            this.updateSettings(settings);
+        });
+
+        colorToggle.on("change",event => {
+            let id = event.target.id.replace('toggle','')-1;
+            let settings = game.settings.get(MODULE.moduleName,'soundboardSettings');
+            settings.toggle[id]=event.target.value;
+            this.updateSettings(settings);
+        });
 
         colorPickerOn.on('click',(event) => {
             const target = event.currentTarget.value;
             let color = document.getElementById("colorOn"+target).value;
             if ((color < 0 && color > 127) || color == "") color = 0;
-            launchpad.colorPicker(target,1,color);
-            //launchpad.colorPicker(0);  
+            launchpad.colorPicker(target,1,color); 
         });
+
+        colorPickerOnNr.on('change',(event) => {
+            let id = event.target.id.replace('colorOn','')-1;
+            let settings = game.settings.get(MODULE.moduleName,'soundboardSettings');
+            settings.colorOn[id]=event.target.value;
+            this.updateSettings(settings);
+        });
+
         colorPickerOff.on('click',(event) => {
             const target = event.currentTarget.value;
             let color = document.getElementById("colorOff"+target).value;
@@ -190,46 +295,90 @@ export class soundboardConfigForm extends FormApplication {
             launchpad.colorPicker(target,0,color);
             
         });
-        if (playlistSelect.length > 0) {
-            playlistSelect.on("change", event => {
-                this.playlist = event.target.value;
-                let settings = game.settings.get(MODULE.moduleName,'soundboardSettings');
-                settings.playlist = this.playlist;
-                game.settings.set(MODULE.moduleName,'soundboardSettings');
-                this.updatePlaylist = true;
-                this.render();
-            });
-        }
-        volumeSlider.on('change', event => {
-            let id = event.target.id.replace('volume','');
-            const column = id%10-1;
-            const row = 8-Math.floor(id/10);
-            id = row*8+column;
+
+        colorPickerOffNr.on('change',(event) => {
+            let id = event.target.id.replace('colorOff','')-1;
             let settings = game.settings.get(MODULE.moduleName,'soundboardSettings');
-            settings.volume[id] = event.target.value;
-            game.settings.set(MODULE.moduleName,'soundboardSettings',settings);
-            if (soundboard.activeSounds[id] != false){
-                const volume = AudioHelper.inputToVolume(event.target.value/100) * game.settings.get("core", "globalInterfaceVolume");
-                soundboard.activeSounds[id].volume(volume);
-            }
+            settings.colorOff[id]=event.target.value;
+            this.updateSettings(settings);
         });
-        if (soundSelect.length > 0) {
-            soundSelect.on("change",event => {
-                let id = event.target.id.replace('soundSelect','');
-                const column = id%10-1;
-                const row = 8-Math.floor(id/10);
-                id = row*8+column;
-                let settings = game.settings.get(MODULE.moduleName,'soundboardSettings');
-                settings.sounds[id] = event.target.value;
-                game.settings.set(MODULE.moduleName,'soundboardSettings',settings);
-                if (soundboard.activeSounds[id] != false){
-                    const mode = settings.mode[id];
-                    let repeat = false;
-                    if (mode == 1) repeat = true;
-                    soundboard.playSound(id,repeat,false);
-                }
-            });
+        
+        playMode.on("change",event => {
+            let id = event.target.id.replace('playmode','')-1;
+            let settings = game.settings.get(MODULE.moduleName,'soundboardSettings');
+            settings.mode[id]=event.target.value;
+            this.updateSettings(settings);
+        });
+
+        volumeSlider.on('change', event => {
+            let id = event.target.id.replace('volume','')-1;
+            let settings = game.settings.get(MODULE.moduleName,'soundboardSettings');
+            settings.volume[id]=event.target.value;
+            this.updateSettings(settings); 
+        });
+        
+        clearAll.on('click',(event) => {
+            let d = new Dialog({
+                title: game.i18n.localize("MaterialKeys.ClearData"),
+                content: "<p>"+game.i18n.localize("MaterialKeys.ClearDataContent")+"</p>",
+                buttons: {
+                 yes: {
+                  icon: '<i class="fas fa-check"></i>',
+                  label: game.i18n.localize("MaterialKeys.Yes"),
+                  callback: () => this.updateSettings('clear',true)
+                 },
+                 no: {
+                  icon: '<i class="fas fa-times"></i>',
+                  label: game.i18n.localize("MaterialKeys.No")
+                 }
+                },
+                default: "no"
+               });
+               d.render(true);
+            this.render();
+        });
+    }
+    async updateSettings(settings,render=false){
+        if (settings == 'clear'){
+            console.log("Clearing soundboard configuration")
+            let colorOff = [];
+            let colorOn = [];
+            let mode = [];
+            let name = [];
+            let selectedPlaylists = [];
+            let sounds = [];
+            let src = [];
+            let toggle = [];
+            let volume = [];
+            for (let i=0; i<64; i++){
+                colorOff[i] = "";
+                colorOn[i] = "";
+                mode[i] = 0;
+                name[i] = "";
+                selectedPlaylists[i] = "none";
+                sounds[i] = "";
+                src[i] = "";
+                toggle[i] = 0;
+                volume[i] = 50;
+            }
+            settings = {
+                colorOff: colorOff,
+                colorOn: colorOn,
+                mode: mode,
+                name: name,
+                selectedPlaylists: selectedPlaylists,
+                sounds: sounds,
+                src: src,
+                toggle: toggle,
+                volume: volume
+            };
         }
+        await game.settings.set(MODULE.moduleName,'soundboardSettings',settings);
+        if (MODULE.enableModule) {
+            launchpad.setMode(MODULE.launchpad.keyMode);
+            soundboard.update();
+        }
+        if (render) this.render();
     }
     
 }
@@ -260,44 +409,42 @@ export class macroConfigForm extends FormApplication {
      * Provide data to the template
      */
     getData() {
-        var selectedMacros = game.settings.get(MODULE.moduleName,'macroSettings').macros;
-        var color = game.settings.get(MODULE.moduleName,'macroSettings').color;
-        var args = game.settings.get(MODULE.moduleName,'macroArgs');
-        if (selectedMacros == undefined) selectedMacros = [];
-        if (color == undefined) color = [];
-        if (args == undefined) args = [];
-        let macroData = {};
+        let settings = game.settings.get(MODULE.moduleName,'macroSettings');
+        if (settings.macros == undefined) settings.macros = [];
+        if (settings.color == undefined) settings.color = [];
+        if (settings.args == undefined) settings.args = [];
+        let macroData = [];
 
         let furnaceEnabled = false;
         const furnace = game.modules.get("furnace");
         if (furnace != undefined && furnace.active) furnaceEnabled = true;
-        let height = 100;
+        let height = 95;
         if (furnaceEnabled) height += 50;
 
+        let iteration = 0;
         for (let j=0; j<8; j++){
-            let macroThis = {};
+            let macroThis = [];
       
             for (let i=0; i<8; i++){
                 const dataThis = {
-                    iteration: 10*(8-j)+i+1,
-                    macro: selectedMacros[j*8+i],
-                    color: color[j*8+i],
+                    iteration: iteration+1,
+                    macro: settings.macros[iteration],
+                    color: settings.color[iteration],
                     macros:game.macros,
-                    args: args[j*8+i],
+                    args: settings.args[iteration],
                     furnace: furnaceEnabled
                 }
-                MODULE.setToJSONArray(macroThis,i,dataThis);
+                macroThis.push(dataThis);
+                iteration++;
             }
             const data = {
                 dataThis: macroThis,
             };
-            MODULE.setToJSONArray(macroData,j,data);
+            macroData.push(data);
         }
-        
+
         return {
             height: height,
-            macros: game.macros,
-            selectedMacros: selectedMacros,
             macroData: macroData,
         } 
     }
@@ -323,7 +470,25 @@ export class macroConfigForm extends FormApplication {
 
     activateListeners(html) {
         super.activateListeners(html);
+        const macro = html.find("select[name='macros']");
+        const args = html.find("input[name='args']");
         const colorPicker = html.find("button[name='colorPicker']");
+        const colorPickerNr = html.find("input[name='color']");
+        const clearAll = html.find("button[name='clearAll']")
+
+        macro.on("change", event => {
+            let id = event.target.id.replace('macros','');
+            let settings = game.settings.get(MODULE.moduleName,'macroSettings');
+            settings.macros[id-1]=event.target.value;
+            this.updateSettings(settings);
+        });
+
+        args.on("change", event => {
+            let id = event.target.id.replace('args','');
+            let settings = game.settings.get(MODULE.moduleName,'macroSettings');
+            settings.args[id-1]=event.target.value;
+            this.updateSettings(settings);
+        });
 
         colorPicker.on('click',(event) => {
             const target = event.currentTarget.value;
@@ -331,5 +496,57 @@ export class macroConfigForm extends FormApplication {
             if ((color < 0 && color > 127) || color == "") color = 0;
             launchpad.colorPicker(target,0,color);
         });
+
+        colorPickerNr.on('change',(event) => {
+            let id = event.target.id.replace('color','')-1;
+            let j = Math.floor(id/8);
+            let i = id % 8;
+            let settings = game.settings.get(MODULE.moduleName,'macroSettings');
+            settings.color[id]=event.target.value;
+            this.updateSettings(settings);
+        });
+
+        clearAll.on('click',(event) => {
+            let d = new Dialog({
+                title: game.i18n.localize("MaterialKeys.ClearData"),
+                content: "<p>"+game.i18n.localize("MaterialKeys.ClearDataContent")+"</p>",
+                buttons: {
+                 yes: {
+                  icon: '<i class="fas fa-check"></i>',
+                  label: game.i18n.localize("MaterialKeys.Yes"),
+                  callback: () => this.updateSettings('clear',true)
+                 },
+                 no: {
+                  icon: '<i class="fas fa-times"></i>',
+                  label: game.i18n.localize("MaterialKeys.No")
+                 }
+                },
+                default: "no"
+               });
+               d.render(true);
+            this.render();
+        });
+    }
+
+    async updateSettings(settings,render=false){
+        if (settings == 'clear'){
+            console.log("Clearing macro board configuration")
+            let color = [];
+            let args = [];
+            let macros = [];
+            for (let i=0; i<64; i++){
+                color[i] = 0;
+                args[i] = "";
+                macros[i] = "none";
+            }
+            settings = {
+                color: color,
+                args: args,
+                macros: macros
+            };
+        }
+        await game.settings.set(MODULE.moduleName,'macroSettings',settings);
+        if (MODULE.enableModule) macroBoard.update();
+        if (render) this.render();
     }
 }